move component to folder, fix issue in data file
