+++ src/components/contentComponent.js
